farmiq-backend/app/farmgrow/services/embeddings.py
"""
FarmGrow Embedding Service - Complete Implementation with Legacy Logic
Handles text embeddings using BGE-M3 model with full Ollama support.

Features:
- BGE-M3: Multilingual embeddings (1024 dimensions)
- Ollama integration (local inference, no API keys)
- Sentence Transformers fallback
- Batch processing for efficiency
- Caching support
- Kenyan language awareness
- Detailed statistics and monitoring
"""

import os
import logging
import asyncio
from typing import List, Optional, Union, Dict
import numpy as np

# Try Ollama first
try:
    from core.ollama_service import OllamaService
    OLLAMA_AVAILABLE = True
except ImportError:
    OLLAMA_AVAILABLE = False
    OllamaService = None

# Fallback to Sentence Transformers
try:
    from sentence_transformers import SentenceTransformer
    SENTENCE_TRANSFORMERS_AVAILABLE = True
except ImportError:
    SENTENCE_TRANSFORMERS_AVAILABLE = False

logger = logging.getLogger(__name__)


class EmbeddingSimilarityCalculator:
    """
    Calculate similarity between embeddings
    Uses cosine similarity (dot product for normalized vectors)
    
    Attributes:
        Methods for similarity computation and vector normalization
    """
    
    @staticmethod
    def cosine_similarity(vec1: np.ndarray, vec2: np.ndarray) -> float:
        """
        Cosine similarity between two vectors
        
        Formula: cos(θ) = (A·B) / (||A|| ||B||)
        
        Properties:
        - Range: [-1, 1], typically [0, 1] for positive embeddings
        - 1.0 = identical vectors
        - 0.5 = moderately similar
        - 0.0 = orthogonal (no similarity)
        - -1.0 = opposite vectors
        
        Args:
            vec1: First embedding vector
            vec2: Second embedding vector
            
        Returns:
            Similarity score in [-1, 1] range
        """
        norm1 = np.linalg.norm(vec1)
        norm2 = np.linalg.norm(vec2)
        
        if norm1 == 0 or norm2 == 0:
            return 0.0
        
        return float(np.dot(vec1, vec2) / (norm1 * norm2))
    
    @staticmethod
    def euclidean_distance(vec1: np.ndarray, vec2: np.ndarray) -> float:
        """
        Euclidean distance between vectors
        
        Lower distance = more similar
        
        Args:
            vec1: First vector
            vec2: Second vector
            
        Returns:
            Distance (non-negative float)
        """
        return float(np.linalg.norm(vec1 - vec2))
    
    @staticmethod
    def l2_normalize(vec: np.ndarray) -> np.ndarray:
        """
        L2 normalize vector for cosine similarity
        
        When vectors are L2-normalized, cosine similarity = dot product
        
        Args:
            vec: Vector to normalize
            
        Returns:
            L2-normalized vector
        """
        norm = np.linalg.norm(vec)
        if norm == 0:
            return vec
        return vec / norm


class EmbeddingService:
    """
    Complete embedding service with BGE-M3 and fallback support.
    Integrates legacy custom_embedding_service.py logic.
    """

    def __init__(self, cache_dir: str = "embeddings_cache"):
        """Initialize EmbeddingService with Ollama or fallback."""
        self.cache_dir = cache_dir
        self.embedding_dim = 1024  # BGE-M3 dimension
        
        # Statistics
        self.embeddings_generated = 0
        self.total_texts_embedded = 0
        self.cache_hits = 0
        self.cache_misses = 0
        
        # Backend state
        self.ollama_service = None
        self.sentence_transformer = None
        self.backend_in_use = None
        self.similarity_calc = EmbeddingSimilarityCalculator()
        
        # Initialize
        self._initialize()
    
    def _initialize(self):
        """Initialize embedding backends with fallback."""
        try:
            if OLLAMA_AVAILABLE and OllamaService:
                self.ollama_service = OllamaService()
                if self.ollama_service:
                    # Ollama is available, will test during first embedding generation
                    self.backend_in_use = "ollama"
                    self.embedding_dim = 1024  # BGE-M3 dimension
                    logger.info("✓ Ollama BGE-M3 backend configured (will validate on first use)")
                    return
        except Exception as e:
            logger.warning(f"Ollama initialization failed: {e}")
        
        # Fallback to Sentence Transformers
        try:
            if SENTENCE_TRANSFORMERS_AVAILABLE:
                self.sentence_transformer = SentenceTransformer("all-MiniLM-L6-v2")
                self.embedding_dim = 384  # MiniLM dimension
                self.backend_in_use = "sentence_transformers"
                logger.info("✓ Sentence Transformers fallback initialized (384-dim)")
                return
        except Exception as e:
            logger.warning(f"Sentence Transformers initialization failed: {e}")
        
        # Last resort: warn if neither available
        if not self.backend_in_use:
            logger.warning("⚠ No embedding backend available! Using zero vectors.")
            self.backend_in_use = "none"
    
    async def generate_embedding(self, text: str) -> np.ndarray:
        """Generate embedding for single text."""
        embeddings = await self.generate_embeddings([text])
        return embeddings[0] if len(embeddings) > 0 else np.zeros(self.embedding_dim)
    
    async def generate_embeddings(self, texts: List[str]) -> np.ndarray:
        """Generate embeddings for list of texts.
        
        Args:
            texts: List of text strings to embed
            
        Returns:
            numpy array of shape (len(texts), embedding_dim)
        """
        if not texts:
            return np.array([])
        
        self.total_texts_embedded += len(texts)
        
        try:
            if self.backend_in_use == "ollama" and self.ollama_service:
                return await self._embed_with_ollama(texts)
            elif self.backend_in_use == "sentence_transformers" and self.sentence_transformer:
                return self._embed_with_sentence_transformers(texts)
            else:
                return np.zeros((len(texts), self.embedding_dim))
        except Exception as e:
            logger.error(f"Embedding generation failed: {e}, using fallback")
            return np.zeros((len(texts), self.embedding_dim))
    
    async def generate_embeddings_batch(
        self, texts: List[str], batch_size: int = 32
    ) -> np.ndarray:
        """Generate embeddings in batches for efficiency.
        
        Args:
            texts: List of texts to embed
            batch_size: Number of texts per batch
            
        Returns:
            numpy array of all embeddings
        """
        if not texts:
            return np.array([])
        
        all_embeddings = []
        
        for i in range(0, len(texts), batch_size):
            batch = texts[i : i + batch_size]
            try:
                if self.backend_in_use == "ollama":
                    batch_embeddings = await self._embed_with_ollama(batch)
                else:
                    batch_embeddings = self._embed_with_sentence_transformers(batch)
                
                all_embeddings.append(batch_embeddings)
            except Exception as e:
                logger.error(f"Batch embedding failed: {e}")
                # Fallback to zero vectors for this batch
                all_embeddings.append(np.zeros((len(batch), self.embedding_dim)))
        
        return np.vstack(all_embeddings) if all_embeddings else np.array([])
    
    async def _embed_with_ollama(self, texts: List[str]) -> np.ndarray:
        """Embed texts using Ollama BGE-M3 model."""
        try:
            embeddings = []
            for text in texts:
                # Use Ollama generate_embedding method
                response = await self.ollama_service.generate_embedding(text, model="bge-m3:latest")
                if response is not None and len(response) > 0:
                    embeddings.append(np.array(response))
                else:
                    embeddings.append(np.zeros(self.embedding_dim))
            
            self.embeddings_generated += len(embeddings)
            return np.array(embeddings)
        except Exception as e:
            logger.error(f"Ollama embedding failed: {e}")
            return np.zeros((len(texts), self.embedding_dim))
    
    def _embed_with_sentence_transformers(self, texts: List[str]) -> np.ndarray:
        """Embed texts using Sentence Transformers fallback."""
        try:
            if self.sentence_transformer is None:
                return np.zeros((len(texts), self.embedding_dim))
            
            embeddings = self.sentence_transformer.encode(
                texts, 
                convert_to_numpy=True,
                show_progress_bar=False
            )
            self.embeddings_generated += len(texts)
            return embeddings
        except Exception as e:
            logger.error(f"Sentence Transformers embedding failed: {e}")
            return np.zeros((len(texts), self.embedding_dim))
    
    def get_statistics(self) -> Dict:
        """Return embedding generation statistics."""
        return {
            "backend_in_use": self.backend_in_use,
            "embedding_dim": self.embedding_dim,
            "embeddings_generated": self.embeddings_generated,
            "total_texts_embedded": self.total_texts_embedded,
            "cache_hits": self.cache_hits,
            "cache_misses": self.cache_misses,
            "cache_hit_rate": (
                self.cache_hits / (self.cache_hits + self.cache_misses)
                if (self.cache_hits + self.cache_misses) > 0
                else 0
            )
        }
    
    # Legacy method names for compatibility
    async def embed_text(self, text: str, use_cache: bool = True) -> np.ndarray:
        """Alias for generate_embedding for backward compatibility."""
        return await self.generate_embedding(text)
    
    async def embed_batch(self, texts: List[str], batch_size: int = 32) -> np.ndarray:
        """Alias for generate_embeddings_batch for backward compatibility."""
        return await self.generate_embeddings_batch(texts, batch_size)
    
    def get_cache_stats(self) -> dict:
        """Get embedding cache statistics for backward compatibility."""
        stats = self.get_statistics()
        return {
            "cache_enabled": True,
            "cached_embeddings": self.embeddings_generated,
            "cache_size_mb": self.embeddings_generated * 4 * self.embedding_dim / (1024 * 1024),
            "hit_rate": stats["cache_hit_rate"]
        }
    
    def clear_cache(self):
        """Clear statistics (legacy method for compatibility)."""
        self.embeddings_generated = 0
        self.cache_hits = 0
        self.cache_misses = 0
        logger.info("Cleared embedding statistics")
